+++ src/main/java/terramine/common/components/MovementOrderComponent.java
@@ -5,12 +5,14 @@ import dev.onyxstudios.cca.api.v3.component.sync.AutoSyncedComponent;
 import dev.onyxstudios.cca.api.v3.entity.PlayerComponent;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.player.Player;
+import org.jetbrains.annotations.NotNull;
 
 @SuppressWarnings("UnstableApiUsage")
 public class MovementOrderComponent implements PlayerComponent<Component>, AutoSyncedComponent {
     private final Player provider;
     private boolean cloud;
     private boolean wings;
+    private boolean wallJump;
 
     public MovementOrderComponent(Player provider) {
         this.provider = provider;
@@ -32,11 +34,19 @@ public class MovementOrderComponent implements PlayerComponent<Component>, AutoS
         return wings;
     }
 
-    @Override
-    public void readFromNbt(CompoundTag tag) {
+    public void setWallJumped(boolean wallJump) {
+        this.wallJump = wallJump;
+    }
+
+    public boolean getWallJumped() {
+        return wallJump;
     }
 
     @Override
-    public void writeToNbt(CompoundTag tag) {
+    public void readFromNbt(@NotNull CompoundTag tag) {
+    }
+
+    @Override
+    public void writeToNbt(@NotNull CompoundTag tag) {
     }
 }