+++ src/main/java/terramine/common/entity/FallingStarEntity.java
@@ -1,7 +1,5 @@
 package terramine.common.entity;
 
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.RandomSource;