+++ src/main/java/terramine/mixin/item/bundleofballoons/LivingEntityMixin.java
@@ -58,7 +58,7 @@ public abstract class LivingEntityMixin extends Entity implements LivingEntityEx
 		ClientPlayNetworking.send(BundleOfBalloonsItem.C2S_QUADRUPLE_JUMPED_ID, PacketByteBufs.empty());
 	}
 
-	@ModifyVariable(method = "causeFallDamage", ordinal = 0, at = @At("HEAD"))
+	@ModifyVariable(method = "causeFallDamage", ordinal = 0, at = @At("HEAD"), argsOnly = true)
 	private float reduceFallDistance(float fallDistance) {
 		// FIXME: this probably also works if we didn't double jump, intended?
 		if (TrinketsHelper.isEquipped(ModItems.BUNDLE_OF_BALLOONS, (LivingEntity) (Object) this)) {