+++ src/main/java/terramine/mixin/item/cloudinabottle/LivingEntityMixin.java
@@ -96,7 +96,7 @@ public abstract class LivingEntityMixin extends Entity implements LivingEntityEx
 		ClientPlayNetworking.send(CloudInABottleItem.C2S_DOUBLE_JUMPED_ID, PacketByteBufs.empty());
 	}
 
-	@ModifyVariable(method = "causeFallDamage", ordinal = 0, at = @At("HEAD"))
+	@ModifyVariable(method = "causeFallDamage", ordinal = 0, at = @At("HEAD"), argsOnly = true)
 	private float reduceFallDistance(float fallDistance) {
 		// FIXME: this probably also works if we didn't double jump, intended?
 		if (TrinketsHelper.isEquipped(ModItems.CLOUD_IN_A_BOTTLE, (LivingEntity) (Object) this)) {