+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2022-07-01
+### Added
+- Dungeon Structure
+- Dungeon Blocks
+
+### Changed
+- Rocket Boot type accessories refill after wall jumping
+- Demon Eyes now bounce when hitting blocks
+
 ## [1.3.6] - 2022-06-25
 ### Added
 - Re-added Cobalt and Obsidian shield (no longer uses Fabric Shield API)